# docs/swagger.yaml
definitions:
  commons.SwaggerApiResponseError:
    properties:
      data: {}
      message:
        type: string
      success:
        default: false
        type: boolean
    type: object
  commons.SwaggerApiResponseSuccessWithData:
    properties:
      data: {}
      message:
        type: string
      success:
        default: true
        type: boolean
    type: object
  commons.SwaggerApiResponseSuccessWithoutData:
    properties:
      data: {}
      message:
        type: string
      success:
        default: true
        type: boolean
    type: object
...
  models.ItemRequest:
    properties:
      item_type_id:
        type: integer
      name:
        type: string
      price:
        type: integer
      profit:
        type: integer
    type: object
  models.ItemTypeRequest:
    properties:
      type:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a simple store management server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Simple Store Management API
  version: "1.0"
paths:
...
  /api/branch/items/{id}:
    get:
      consumes:
      - application/json
      description: get branch with Iitems
      parameters:
      - description: Branch ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/commons.SwaggerApiResponseSuccessWithData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/commons.SwaggerApiResponseError'
      security:
      - Bearer: []
      summary: Get Branch With Items
      tags:
      - branch
...
  /api/item:
    get:
      consumes:
      - application/json
      description: get all items
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/commons.SwaggerApiResponseSuccessWithData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/commons.SwaggerApiResponseError'
      security:
      - Bearer: []
      summary: Get All Items
      tags:
      - item
    post:
      consumes:
      - application/json
      description: create a new item
      parameters:
      - description: Item Request
        in: body
        name: itemRequest
        required: true
        schema:
          $ref: '#/definitions/models.ItemRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/commons.SwaggerApiResponseSuccessWithoutData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/commons.SwaggerApiResponseError'
      security:
      - Bearer: []
      summary: Create Item
      tags:
      - item
  /api/item/{id}:
    delete:
      consumes:
      - application/json
      description: delete item by id
      parameters:
      - description: Item ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/commons.SwaggerApiResponseSuccessWithoutData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/commons.SwaggerApiResponseError'
      security:
      - Bearer: []
      summary: Delete Item
      tags:
      - item
    get:
      consumes:
      - application/json
      description: get item by id
      parameters:
      - description: Item ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/commons.SwaggerApiResponseSuccessWithData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/commons.SwaggerApiResponseError'
      security:
      - Bearer: []
      summary: Get Item
      tags:
      - item
    put:
      consumes:
      - application/json
      description: update item by id
      parameters:
      - description: Item Request
        in: body
        name: itemRequest
        required: true
        schema:
          $ref: '#/definitions/models.ItemRequest'
      - description: Item ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/commons.SwaggerApiResponseSuccessWithoutData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/commons.SwaggerApiResponseError'
      security:
      - Bearer: []
      summary: Update Item
      tags:
      - item
  /api/itemType:
    get:
      consumes:
      - application/json
      description: get all items types
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/commons.SwaggerApiResponseSuccessWithData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/commons.SwaggerApiResponseError'
      security:
      - Bearer: []
      summary: Get All Item Types
      tags:
      - itemType
    post:
      consumes:
      - application/json
      description: create a new item type
      parameters:
      - description: Item Type Request
        in: body
        name: itemTypeRequest
        required: true
        schema:
          $ref: '#/definitions/models.ItemTypeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/commons.SwaggerApiResponseSuccessWithoutData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/commons.SwaggerApiResponseError'
      security:
      - Bearer: []
      summary: Create Item Type
      tags:
      - itemType
  /api/itemType/{id}:
    delete:
      consumes:
      - application/json
      description: delete item type by id
      parameters:
      - description: Item Type ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/commons.SwaggerApiResponseSuccessWithoutData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/commons.SwaggerApiResponseError'
      security:
      - Bearer: []
      summary: Delete Item Type
      tags:
      - itemType
    get:
      consumes:
      - application/json
      description: get item type by id
      parameters:
      - description: Item Type ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/commons.SwaggerApiResponseSuccessWithData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/commons.SwaggerApiResponseError'
      security:
      - Bearer: []
      summary: Get Item Type
      tags:
      - itemType
    put:
      consumes:
      - application/json
      description: update item type by id
      parameters:
      - description: Item Type Request
        in: body
        name: itemTypeRequest
        required: true
        schema:
          $ref: '#/definitions/models.ItemTypeRequest'
      - description: Branch Item ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/commons.SwaggerApiResponseSuccessWithoutData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/commons.SwaggerApiResponseError'
      security:
      - Bearer: []
      summary: Update Item Type
      tags:
      - itemType
  /api/itemType/items/{id}:
    get:
      consumes:
      - application/json
      description: get items of item type
      parameters:
      - description: Item Type ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/commons.SwaggerApiResponseSuccessWithData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/commons.SwaggerApiResponseError'
      security:
      - Bearer: []
      summary: Get Items Of Item Type
      tags:
      - itemType
...
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"